pyproject.toml
[project]
name = "vestaboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
    "spotipy>=2.25.1",
    "unidecode>=1.4.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
